+++ dst/app/routes/transactions/placeOrder.js
@@ -432,7 +432,10 @@ placeOrderTransactionsRouter.post('/:transactionId/tasks/sendEmailNotification',
             },
             about: req.body.about,
             text: req.body.text
-        })(new ttts.repository.Task(mongoose.connection), new ttts.repository.Transaction(mongoose.connection));
+        })({
+            task: new ttts.repository.Task(mongoose.connection),
+            transaction: new ttts.repository.Transaction(mongoose.connection)
+        });
         res.status(http_status_1.CREATED)
             .json(task);
     }