+++ dst/jobs/continuous/watchConfirmedTransaction/run.js
@@ -13,7 +13,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
  */
 const ttts = require("@tokyotower/domain");
 const connectMongo_1 = require("../../../connectMongo");
-exports.default = (_) => __awaiter(this, void 0, void 0, function* () {
+exports.default = (params) => __awaiter(this, void 0, void 0, function* () {
     const connection = yield connectMongo_1.connectMongo({ defaultConnection: false });
     let countExecute = 0;
     const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
@@ -26,7 +26,13 @@ exports.default = (_) => __awaiter(this, void 0, void 0, function* () {
         }
         countExecute += 1;
         try {
-            yield ttts.service.transaction.placeOrder.exportTasks(ttts.factory.transactionStatusType.Confirmed)(taskRepo, transactionRepo);
+            yield ttts.service.transaction.placeOrder.exportTasks({
+                project: params.project,
+                status: ttts.factory.transactionStatusType.Confirmed
+            })({
+                task: taskRepo,
+                transaction: transactionRepo
+            });
         }
         catch (error) {
             // tslint:disable-next-line:no-console